# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "everos",
    "description": "md-first memory extraction framework",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/memory/add": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Add Memory",
        "description": "Add messages into the user-memory + agent-memory pipelines.",
        "operationId": "add_memory_api_v1_memory_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorizeAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_AddResponseData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/stage": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Stage Memory",
        "description": "Durably stage messages without invoking boundary or extraction.",
        "operationId": "stage_memory_api_v1_memory_stage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorizeStageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_StageResponseData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/flush": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Flush Memory",
        "description": "Force boundary detection over the current ``session_id`` buffer.\n\n[OSS-only] — cloud edition decides boundary timing server-side and\ndoes not expose this endpoint.",
        "operationId": "flush_memory_api_v1_memory_flush_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorizeFlushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_FlushResponseData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/operations/{operation_id}": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Memory Operation Status",
        "description": "Inspect a write receipt without returning conversation content.",
        "operationId": "memory_operation_status_api_v1_memory_operations__operation_id__get",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_MemoryOperationStatus_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddResponseData": {
        "properties": {
          "message_count": {
            "type": "integer",
            "title": "Message Count"
          },
          "status": {
            "type": "string",
            "enum": [
              "accumulated",
              "extracted"
            ],
            "title": "Status"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "replayed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replayed"
          }
        },
        "type": "object",
        "required": [
          "message_count",
          "status"
        ],
        "title": "AddResponseData"
      },
...
      "ContentItemDTO": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "doc",
              "pdf",
              "html",
              "email"
            ],
            "title": "Type"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uri"
          },
          "base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base64"
          },
          "ext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ext"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "extras": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extras"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "ContentItemDTO",
        "description": "Content piece (v1 API brief appendix A)."
      },
...
      "FlushResponseData": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "extracted",
              "no_extraction",
              "processing"
            ],
            "title": "Status"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "replayed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replayed"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FlushResponseData"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemorizeAddRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id"
          },
          "app_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.@+-]+$",
            "title": "App Id",
            "default": "default"
          },
          "project_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.@+-]+$",
            "title": "Project Id",
            "default": "default"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageItemDTO"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Messages"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^evop1-add-[0-9a-f]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id",
            "description": "Optional caller-stable idempotency key."
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "messages"
        ],
        "title": "MemorizeAddRequest"
      },
      "MemorizeFlushRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id"
          },
          "app_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.@+-]+$",
            "title": "App Id",
            "default": "default"
          },
          "project_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.@+-]+$",
            "title": "Project Id",
            "default": "default"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^evop1-flush-[0-9a-f]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id",
            "description": "Optional caller-stable idempotency key."
          },
          "background": {
            "type": "boolean",
            "title": "Background",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "MemorizeFlushRequest"
      },
      "MemorizeStageRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id"
          },
          "app_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.@+-]+$",
            "title": "App Id",
            "default": "default"
          },
          "project_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.@+-]+$",
            "title": "Project Id",
            "default": "default"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageItemDTO"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Messages"
          },
          "operation_id": {
            "type": "string",
            "pattern": "^evop1-stage-[0-9a-f]{64}$",
            "title": "Operation Id",
            "description": "Caller-stable deferred-ingest idempotency key."
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "messages",
          "operation_id"
        ],
        "title": "MemorizeStageRequest"
      },
      "MemoryOperationStatus": {
        "properties": {
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "add",
              "stage",
              "flush"
            ],
            "title": "Kind"
          },
          "app_id": {
            "type": "string",
            "title": "App Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "failed"
            ],
            "title": "State"
          },
          "stage": {
            "type": "string",
            "enum": [
              "claimed",
              "messages_staged",
              "queued",
              "processing",
              "memcells_committed",
              "sync_dispatch_completed"
            ],
            "title": "Stage"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count"
          },
          "memcell_count": {
            "type": "integer",
            "title": "Memcell Count"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          }
        },
        "type": "object",
        "required": [
          "operation_id",
          "kind",
          "app_id",
          "project_id",
          "session_id",
          "state",
          "stage",
          "message_count",
          "memcell_count",
          "retryable"
        ],
        "title": "MemoryOperationStatus",
        "description": "Safe, content-free view exposed by the operation query endpoint."
      },
...
      "MessageItemDTO": {
        "properties": {
          "sender_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.@+-]+$",
            "title": "Sender Id"
          },
          "sender_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant",
              "tool"
            ],
            "title": "Role"
          },
          "timestamp": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Timestamp",
            "description": "Message event time as Unix epoch in **milliseconds** (v1 API contract; the algo layer auto-detects sec vs ms for backward compat but the contract is ms)."
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ContentItemDTO"
                },
                "type": "array"
              }
            ],
            "title": "Content"
          },
          "tool_calls": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ToolCallDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Calls"
          },
          "tool_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Id"
          }
        },
        "type": "object",
        "required": [
          "sender_id",
          "role",
          "timestamp",
          "content"
        ],
        "title": "MessageItemDTO"
      },
...
      "StageResponseData": {
        "properties": {
          "message_count": {
            "type": "integer",
            "title": "Message Count"
          },
          "status": {
            "type": "string",
            "const": "staged",
            "title": "Status"
          },
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed"
          }
        },
        "type": "object",
        "required": [
          "message_count",
          "status",
          "operation_id",
          "replayed"
        ],
        "title": "StageResponseData"
      },
      "SuccessEnvelope_AddResponseData_": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "data": {
            "$ref": "#/components/schemas/AddResponseData"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "SuccessEnvelope[AddResponseData]"
      },
      "SuccessEnvelope_FlushResponseData_": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "data": {
            "$ref": "#/components/schemas/FlushResponseData"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "SuccessEnvelope[FlushResponseData]"
      },
      "SuccessEnvelope_MemoryOperationStatus_": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "data": {
            "$ref": "#/components/schemas/MemoryOperationStatus"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "SuccessEnvelope[MemoryOperationStatus]"
      },
      "SuccessEnvelope_StageResponseData_": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "data": {
            "$ref": "#/components/schemas/StageResponseData"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "title": "SuccessEnvelope[StageResponseData]"
      },
      "ToolCallDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "function"
          },
          "function": {
            "$ref": "#/components/schemas/ToolFunctionDTO"
          }
        },
        "type": "object",
        "required": [
          "id",
          "function"
        ],
        "title": "ToolCallDTO"
      },
      "ToolFunctionDTO": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "arguments": {
            "type": "string",
            "title": "Arguments"
          }
        },
        "type": "object",
        "required": [
          "name",
          "arguments"
        ],
        "title": "ToolFunctionDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}